requirements.txt
django>=1.4
South>=0.7.5
django-social-auth>=0.7.4
PyGithub>=1.4